fix : list component 세분화, List, Lists 컴포넌트 React.memo 적용
