+++ src/components/List.jsx
@@ -1,23 +1,8 @@
 import React from 'react';
 import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd';
+import Lists from './Lists';
 
-export default function List({ todoData, setTodoData }) {
-
-  const handleClick = (id) => {
-    let newTodoData = todoData.filter((data)=> data.id !== id);
-    setTodoData(newTodoData);
-  };
-
-  const handleCompleteChange = (id) => {
-    let newTodoData = todoData.map(data => {
-      if(data.id === id){
-        data.completed = !data.completed;
-      }
-      return data;
-    })
-
-    setTodoData(newTodoData);
-  };
+const List = React.memo(({ todoData, setTodoData }) => {
 
   const handleEnd = (result) => {
     if(!result.destination) return;
@@ -28,6 +13,7 @@ export default function List({ todoData, setTodoData }) {
     setTodoData(newTodoData);
   };
 
+  console.log('list rendering')
   return (
     <div>
       <DragDropContext onDragEnd={handleEnd}>
@@ -38,39 +24,15 @@ export default function List({ todoData, setTodoData }) {
               todoData.map((data, index) => (
                 <Draggable key={data.id} draggableId={data.id.toString()} index={index}>
                   {(provided, snapshot) => (
-                    <div 
-                      className={`${
-                      snapshot.isDragging ? "bg-gray-400" : "bg-gray-100" 
-                      } flex items-center justify-between w-full px-4 py-1 my-2 text-gray-600 border rounded`} 
-                      key={data.id}
-                      {...provided.draggableProps}
-                      {...provided.dragHandleProps}
-                      ref={provided.innerRef}
-                    >
-                      <div className='item-center'>
-                        <input 
-                          type='checkbox' 
-                          defaultCheckd={data.completed} 
-                          onChange={() => handleCompleteChange(data.id)}
-                        />
-                        <span 
-                          className={
-                            data.completed ? "line-through" : undefined
-                          }
-                        >
-                        {data.title}
-                        </span>
-                      </div>
-                    
-                      <div>
-                        <button 
-                          className='px-4 py-2 float-right' 
-                          onClick={() => handleClick(data.id)}
-                        >
-                          x
-                        </button>        
-                      </div>
-                    </div>
+                   <Lists 
+                    id={data.id}
+                    title={data.title}
+                    completed={data.completed}
+                    todoData={todoData}
+                    setTodoData={setTodoData}
+                    provided={provided}
+                    snapshot={snapshot}
+                   />
                   )}
                 </Draggable>
               ))}
@@ -81,4 +43,6 @@ export default function List({ todoData, setTodoData }) {
       </DragDropContext>   
     </div>
   )
-};
+});
+
+export default List;
